[FLINK-11843] Create AbstractDispatcherLeaderProcess
